requirements.txt
cycler==0.10.0
matplotlib==2.1.0
nibabel==2.2.1
numpy==1.13.3
pymvpa2==2.6.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scipy==1.0.0
six==1.11.0
